# Fixes #23492: Perf regression of calling isFirefoxDefault from main thread