--- a/app/src/main/java/org/mozilla/fenix/settings/SettingsFragment.kt
+++ b/app/src/main/java/org/mozilla/fenix/settings/SettingsFragment.kt
@@ -148,7 +148,7 @@ class SettingsFragment : PreferenceFragmentCompat() {
      * Note: Changing Settings screen before experiment is over requires changing all layouts.
      */
     private fun getPreferenceLayoutId() =
-        if (isDefaultBrowserExperimentBranch() && !isFirefoxDefaultBrowser()) {
+        if (isDefaultBrowserExperimentBranch()) {
             R.xml.preferences_default_browser_experiment
         } else {
             R.xml.preferences
@@ -473,7 +473,7 @@ class SettingsFragment : PreferenceFragmentCompat() {
      */
     private fun getClickListenerForMakeDefaultBrowser(): Preference.OnPreferenceClickListener {
         return Preference.OnPreferenceClickListener {
-            if (isDefaultBrowserExperimentBranch() && !isFirefoxDefaultBrowser()) {
+            if (isDefaultBrowserExperimentBranch()) {
                 requireContext().metrics.track(Event.SetDefaultBrowserSettingsScreenClicked)
             }
             activity?.openSetDefaultBrowserOption()
@@ -605,7 +605,7 @@ class SettingsFragment : PreferenceFragmentCompat() {
     }
 
     private fun isDefaultBrowserExperimentBranch(): Boolean =
-        FxNimbus.features.defaultBrowserMessage.value().messageLocation == MessageSurfaceId.SETTINGS
+        requireContext().settings().isDefaultBrowserMessageLocation(MessageSurfaceId.SETTINGS)
 
     private fun isFirefoxDefaultBrowser(): Boolean {
         val browsers = BrowsersCache.all(requireContext())