--- a/nimbus.fml.yaml
+++ b/nimbus.fml.yaml
@@ -54,6 +54,13 @@ features:
         description: If true, the feature shows up on the homescreen and on the new tab screen.
         type: Boolean
         default: false
+    defaults:
+      - channel: nightly
+        value:
+          enabled: true
+      - channel: developer
+        value:
+          enabled: true
   default-browser-message:
     description: A small feature allowing experiments on the placement of a default browser message.
     variables: