Fixes #23492
Fixes #23618

Perf regression calling slow isFirefoxDefault from the main thread.

Prior #23400, the default browser message was only displayed when part of an experiment. The check that isFirefoxDefault was only done as a final check before the message is displayed.

#23400 changed the order of this, so the slow check was done each time the toolbar was created.

Fixes #23492; this PR does three things:

- it moves the check for isDefaultBrowser until _after_ the check to see which message surface to display on (as was prior to #23400 )
- it makes the default browser message be able to turn off.
- it changes the feature of showing the default browser message to be off by default (as was prior to #23400 )

Now, the default browser message is displayable when an experiment toggles it on.

The isFirefoxDefault _is_ done when the message is toggled on, but:

1. There are no plans to toggle this on in this release
2. The messaging system is being re-written to avoid these problems.


### Pull Request checklist
<!-- Before submitting the PR, please address each item -->
- [ ] **Tests**: This PR includes thorough tests or an explanation of why it does not
- [ ] **Screenshots**: This PR includes screenshots or GIFs of the changes made or an explanation of why it does not
- [ ] **Accessibility**: The code in this PR follows [accessibility best practices](https://github.com/mozilla-mobile/shared-docs/blob/master/android/accessibility_guide.md) or does not include any user facing features. In addition, it includes a screenshot of a successful [accessibility scan](https://play.google.com/store/apps/details?id=com.google.android.apps.accessibility.auditor&hl=en_US) to ensure no new defects are added to the product.

### To download an APK when reviewing a PR:
1. click on Show All Checks,
2. click Details next to "Taskcluster (pull_request)" after it appears and then finishes with a green checkmark,
3. click on the "Fenix - assemble" task, then click "Run Artifacts".
4. the APK links should be on the left side of the screen, named for each CPU architecture