@mcomella 2022-02-03
I measured the performance difference. Results from original bisection:
- before regression: 1373ms median
- regression: 1501ms

Results from this PR:
- before PR: 1436ms
- after PR: 1402ms

It seems something may have changed after the initial regression causing the overall regression to be 63ms rather than ~100ms. **One notable visual difference is that in the regressing builds there is a new UI element, "Set as default browser".** The home screen races various UI elements against each other (when it should inflate from the top) so that may contribute to the different results. This PR improves the regression by another 34ms so we've still regressed ~29ms from the original state. Ideally we can try to address that (e.g. if we regress 30ms 2 more times, we're back to where we started) but I didn't see anything obviously actionable in the profiles. I'd be curious to see what our Nightly results say about this PR rather than my local measurements.