@Amejia481 2022-02-07
> One notable visual difference is that in the regressing builds there is a new UI element, "Set as default browser".

@jhugman this looks similar to what @mcomella described on https://github.com/mozilla-mobile/fenix/pull/23556#issuecomment-1029338172 see https://github.com/mozilla-mobile/fenix/issues/23618#issuecomment-1031970923