@Amejia481 2022-02-07
> > One notable visual difference is that in the regressing builds there is a new UI element, "Set as default browser".
> 
> @jhugman this looks similar to what @mcomella described on [#23556 (comment)](https://github.com/mozilla-mobile/fenix/pull/23556#issuecomment-1029338172) see [#23618 (comment)](https://github.com/mozilla-mobile/fenix/issues/23618#issuecomment-1031970923)

I think what is causing the regression is what @mcomella was describing above the **"Set as default browser"** card is showing all the time. Investigating https://github.com/mozilla-mobile/fenix/issues/23618 I found that 
[val isExperimentBranch = feature.messageLocation == MessageSurfaceId.HOMESCREEN_BANNER](https://github.com/mozilla-mobile/fenix/blob/26c7562e62359fa0ec57e8d4f9db1695bb58b45f/app/src/main/java/org/mozilla/fenix/utils/Settings.kt#L331)  it's always true after https://github.com/mozilla-mobile/fenix/commit/82a6f8cae411603ba5abced3b13fc43ca40f335e. See https://github.com/mozilla-mobile/fenix/issues/23618#issuecomment-1032011656 for more context.