@mcomella 2022-02-08
> @mcomella please could you re-profile? (are there profiles on CI for PRs with the Performance label?)

I tried to check out and run this PR but it crashes on start up with:
```
02-07 16:39:03.891 19530 19563 E AndroidRuntime: FATAL EXCEPTION: DefaultDispatcher-worker-5
02-07 16:39:03.891 19530 19563 E AndroidRuntime: Process: org.mozilla.fenix, PID: 19530
02-07 16:39:03.891 19530 19563 E AndroidRuntime: org.mozilla.experiments.nimbus.internal.NimbusFeatureException: A Context is needed but not available. Consider passing in a context to the value() method when close to startup
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at org.mozilla.experiments.nimbus.internal.FeatureHolder.value$default(FeatureHolder.kt:7)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at org.mozilla.fenix.utils.Settings$searchTermTabGroupsAreEnabled$2.invoke(Settings.kt:4)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at org.mozilla.fenix.components.settings.LazyPreference$property$2.invoke(FeatureFlagPreference.kt:4)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at kotlin.SynchronizedLazyImpl.getValue(LazyJVM.kt:5)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at org.mozilla.fenix.components.settings.LazyPreference.getValue(FeatureFlagPreference.kt:3)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at org.mozilla.fenix.utils.Settings.getSearchTermTabGroupsAreEnabled(Settings.kt:1)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at org.mozilla.fenix.FenixApplication$initializeGlean$2.invokeSuspend(FenixApplication.kt:23)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at kotlin.coroutines.jvm.internal.BaseContinuationImpl.resumeWith(ContinuationImpl.kt:3)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at kotlinx.coroutines.DispatchedTask.run(DispatchedTask.kt:18)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at kotlinx.coroutines.scheduling.CoroutineScheduler.runSafely(CoroutineScheduler.kt:1)
02-07 16:39:03.891 19530 19563 E AndroidRuntime:        at kotlinx.coroutines.scheduling.CoroutineScheduler$Worker.run(CoroutineScheduler.kt:10)
```

@jhugman btw, did you mean a benchmark or a profile?

> (are there profiles on CI for PRs with the Performance label?)

Unfortunately, no – we haven't had the resources to implement something like that (e.g. our daily graphs from nightly are actually manually run with a script on a device on my desk).

For benchmarks, you can run it locally if you have a low-end phone (or a high-end phone you can reproduce the regression on). See https://wiki.mozilla.org/Performance/Fenix/Performance_reviews#Benchmark_locally. This regression was to the `cold_main_first_frame` benchmark.