@jhugman 2022-02-08
Ah. Ok yes, I keep forgetting to say: The default browser message was shown as part of the #23400 , and this PR fixes that.

Fixed this early startup crash, with the suggested fix.

```
 org.mozilla.experiments.nimbus.internal.NimbusFeatureException: A Context is needed but not available. Consider passing in a context to the value() method when close to startup
        …
        at org.mozilla.fenix.utils.Settings$searchTermTabGroupsAreEnabled$2.invoke(Settings.kt:4)
```