@jhugman 2022-02-08
> @jhugman btw, did you mean a benchmark or a profile?

I'm not sure, whichever you needed to find that 102ms regression.

Also question from the curious: given the precision of `ms` you're giving, I assumed that the error bars were small, however, knowing you're benchmarking on a local machine, I have no idea now. What sort of variation/error bars do you see in timings?