@mcomella 2022-02-08
> > @jhugman btw, did you mean a benchmark or a profile?
> 
> I'm not sure, whichever you needed to find that 102ms regression.

Sure. Results from original bisection:
- before regression: 1373ms median
- regression: 1501ms

Results from this PR, taken with our `cold_main_first_frame` benchmark run on a Moto G5 (i.e. `./perf-tools/measure_start_up.py nightly cold_main_first_frame results.txt`):
- before PR: 1443ms
- after PR: 1406ms

These results are similar to [when I last measured this PR](https://github.com/mozilla-mobile/fenix/pull/23556#issuecomment-1029338172).

> Also question from the curious: given the precision of `ms` you're giving, I assumed that the error bars were small, however, knowing you're benchmarking on a local machine, I have no idea now. What sort of variation/error bars do you see in timings?

I don't know statistics well so I'm not sure how well I can answer this question 😓 (and yes, our results could probably be improved by someone more familiar with statistics). Through trial-and-error we decided to execute 25 iterations of each test and to take the median values of the results to find the duration of a benchmark for a build. When we originally set these values, this had a maximum variation of ~20ms for an A/A test. In the worst case, I think this would let us catch regressions of ~40ms+ (based [on this thinking I did](https://docs.google.com/document/d/1QKHK50VJisT2k5kTfbNwDb5MvKKTi5H9xAKR9s-Hzu8/edit)). However, the average variation is usually smaller and we can look at the results across multiple days (so hundreds of iterations instead of 25) so in practice we've caught smaller regressions. The variation seems larger these days though:  maybe ~30ms. Our [nightly benchmark results dashboards (populated once a week on mondays) is here](https://earthangel-b40313e5.influxcloud.net/d/s3AT6t7nk/fenix-startup-nightly-via-backfill-py?orgId=1&from=now-30d&to=now) if you're curious.

For context, we use local machines right now rather than CI due to limited resourcing: CI would be ideal because it's more accessible but it takes more time to implement and debug when issues occur.